refactor: extract spending limit table
